Multibody

Documentation for Multibody.

Multibody.world

Multibody.Body

Multibody.BodyShape

Multibody.Damper

Multibody.Prismatic

Multibody.Revolute

Multibody.Spring

Multibody.abs_rotation

Multibody.at_variables_t

Multibody.axisRotation

Multibody.gravity_acceleration

Multibody.resolve1

Multibody.resolve2

Multibody.rotx

Multibody.roty

Multibody.rotz

Multibody.world — Constant

The world component is the root of all multibody models. It is a fixed frame with a parallel gravitational field and a gravity vector specified by the unit direction world.n (defaults to [0, -1, 0]) and magnitude world.g (defaults to 9.81).

Multibody.Body — Method

Body(; name, m = 1, r_cm, I = collect(0.001 * LinearAlgebra.I(3)), isroot = false, phi0 = zeros(3), phid0 = zeros(3), r_0=zeros(3))

Representing a body with 3 translational and 3 rotational degrees-of-freedom.

m: Mass

r_cm: Vector from frame_a to center of mass, resolved in frame_a

I: Inertia matrix of the body

isroot: Indicate whether this component is the root of the system, useful when there are no joints in the model.

phi0: Initial orientation, only applicable if isroot = true

phid0: Initial angular velocity

Multibody.BodyShape — Method

BodyShape(; name, m = 1, r, kwargs...)

The BodyShape component is similar to a Body, but it has two frames and a vector r that describes the translation between them, while the body has a single frame only.

r: Vector from frame_a to frame_b resolved in frame_a

All kwargs are passed to the internal Body component.

Multibody.Damper — Method

Damper(; d, name, kwargs)

Linear damper acting as line force between frame_a and frame_b. A force f is exerted on the origin of frame_b and with opposite sign on the origin of frame_a along the line from the origin of frame_a to the origin of frame_b according to the equation:

\[f = d D(s)\]

where d is the (viscous) damping parameter, s is the distance between the origin of frame_a and the origin of frame_b and D(s) is the time derivative of s.

Arguments:

d: Damping coefficient

Multibody.Prismatic — Method

Prismatic(; name, n = [0, 0, 1], useAxisFlange = false, isroot = true)

Prismatic joint with 1 translational degree-of-freedom

n: The axis of motion (unit vector)

useAxisFlange: If true, the joint will have two additional frames from Mechanical.Translational, axis and support, between which translational components such as springs and dampers can be connected.

isroot: If true, the joint will be considered the root of the system.

If useAxisFlange, flange connectors for ModelicaStandardLibrary.Mechanics.TranslationalModelica are also available:

axis: 1-dim. translational flange that drives the joint

support: 1-dim. translational flange of the drive support (assumed to be fixed in the world frame, NOT in the joint)

The function returns an ODESystem representing the prismatic joint.

Multibody.Revolute — Method

Revolute(; name, phi0 = 0, w0 = 0, n, useAxisFlange = false)

Revolute joint with 1 rotational degree-of-freedom

phi0: Initial angle

w0: Iniitial angular velocity

n: The axis of rotation

useAxisFlange: If true, the joint will have two additional frames from Mechanical.Rotational, axis and support, between which rotational components such as springs and dampers can be connected.

If useAxisFlange, flange connectors for ModelicaStandardLibrary.Mechanics.Rotational are also available:

axis: 1-dim. rotational flange that drives the joint

support: 1-dim. rotational flange of the drive support (assumed to be fixed in the world frame, NOT in the joint)

Multibody.Spring — Method

Spring(; c, name, m = 0, lengthFraction = 0.5, s_unstretched = 0, kwargs)

Linear spring acting as line force between frame_a and frame_b. A force f is exerted on the origin of frame_b and with opposite sign on the origin of frame_a along the line from the origin of frame_a to the origin of frame_b according to the equation:

\[f = c s\]

where c is the spring stiffness parameter, s is the distance between the origin of frame_a and the origin of frame_b.

Optionally, the mass of the spring is taken into account by a point mass located on the line between frame_a and frame_b (default: middle of the line). If the spring mass is zero, the additional equations to handle the mass are removed.

Arguments:

c: Spring stiffness

m: Mass of the spring (can be zero)

lengthFraction: Location of spring mass with respect to frame_a as a fraction of the distance from frame_a to frame_b (=0: at frame_a; =1: at frame_b)

s_unstretched: Length of the spring when it is unstretched

kwargs: are passed to LineForceWithMass

Multibody.abs_rotation — Method

R2 = abs_rotation(R1, R_rel)

R1: Orientation object to rotate frame 0 into frame 1

R_rel: Orientation object to rotate frame 1 into frame 2

R2: Orientation object to rotate frame 0 into frame 2

Multibody.at_variables_t — Method

at_variables_t(args)

Emulates the @variables macro but does never creates array variables. Also never introuces the variable names into the calling scope.

Multibody.axisRotation — Method

axisRotation(sequence, angle; name = :R)

Generate a rotation matrix for a rotation around the specified axis.

sequence: The axis to rotate around (1: x-axis, 2: y-axis, 3: z-axis)

angle: The angle of rotation (in radians)

Returns a RotationMatrix object.

Multibody.gravity_acceleration — Method

Compute the gravity acceleration, resolved in world frame

Multibody.resolve1 — Method

h1 = resolve1(R21, h2)

R12 is a 3x3 matrix that transforms a vector from frame 1 to frame 2. h2 is a vector resolved in frame 2. h1 is the same vector in frame 1.

Typical usage:

resolve1(ori(frame_a), r_ab)

Multibody.resolve2 — Method

h2 = resolve2(R21, h1)

R21 is a 3x3 matrix that transforms a vector from frame 1 to frame 2. h1 is a vector resolved in frame 1. h2 is the same vector in frame 2.

Typical usage:

resolve2(ori(frame_a), a_0 - g_0)

Multibody.rotx — Function

rotx(t, deg = false)

Generate a rotation matrix for a rotation around the x-axis.

t: The angle of rotation (in radians, unless deg is set to true)

deg: (Optional) If true, the angle is in degrees

Returns a 3x3 rotation matrix.

Multibody.roty — Function

roty(t, deg = false)

Generate a rotation matrix for a rotation around the y-axis.

t: The angle of rotation (in radians, unless deg is set to true)

deg: (Optional) If true, the angle is in degrees

Returns a 3x3 rotation matrix.

Multibody.rotz — Function

rotz(t, deg = false)

Generate a rotation matrix for a rotation around the z-axis.

t: The angle of rotation (in radians, unless deg is set to true)

deg: (Optional) If true, the angle is in degrees

Returns a 3x3 rotation matrix.