Frames

Docstrings

Multibody.BodyBox

Multibody.BodyCylinder

Multibody.Planar

Multibody.PlanarMechanics.BodyShape

Multibody.PlanarMechanics.Damper

Multibody.PlanarMechanics.DifferentialGear

Multibody.PlanarMechanics.Fixed

Multibody.PlanarMechanics.FixedTranslation

Multibody.PlanarMechanics.IdealPlanetary

Multibody.PlanarMechanics.PartialAbsoluteBaseSensor

Multibody.PlanarMechanics.PartialAbsoluteSensor

Multibody.PlanarMechanics.PartialRelativeBaseSensor

Multibody.PlanarMechanics.PartialRelativeSensor

Multibody.PlanarMechanics.PartialTwoFrames

Multibody.PlanarMechanics.SimpleWheel

Multibody.PlanarMechanics.Spring

Multibody.PlanarMechanics.SpringDamper

Multibody.PlanarMechanics.ZeroPosition

Multibody.world

Multibody.RotationMatrix

Multibody.AccSensor

Multibody.AxisControlBus

Multibody.AxisType2

Multibody.BasicTorque

Multibody.Body

Multibody.BodyShape

Multibody.CutForce

Multibody.CutTorque

Multibody.Damper

Multibody.Fixed

Multibody.FixedRotation

Multibody.FixedTranslation

Multibody.Force

Multibody.Frame

Multibody.FreeMotion

Multibody.GearConstraint

Multibody.JointRRR

Multibody.JointUSR

Multibody.Kinematic5

Multibody.KinematicPTP

Multibody.NumRotationMatrix

Multibody.PartialCutForceBaseSensor

Multibody.PathPlanning1

Multibody.PathToAxisControlBus

Multibody.PlanarMechanics.AbsolutePosition

Multibody.PlanarMechanics.BasicAbsolutePosition

Multibody.PlanarMechanics.BasicRelativePosition

Multibody.PlanarMechanics.Body

Multibody.PlanarMechanics.Frame

Multibody.PlanarMechanics.Prismatic

Multibody.PlanarMechanics.RelativePosition

Multibody.PlanarMechanics.Revolute

Multibody.PlanarMechanics.SlipBasedWheelJoint

Multibody.PlanarMechanics.limit_S_form

Multibody.PlanarMechanics.limit_S_triple

Multibody.Pose

Multibody.Position

Multibody.Power

Multibody.Prismatic

Multibody.PrismaticConstraint

Multibody.RealPassThrough

Multibody.Revolute

Multibody.RevolutePlanarLoopConstraint

Multibody.RollingConstraintVerticalWheel

Multibody.RollingWheel

Multibody.RollingWheelJoint

Multibody.RollingWheelSet

Multibody.RollingWheelSetJoint

Multibody.Rope

Multibody.SlipWheelJoint

Multibody.SlippingWheel

Multibody.Spherical

Multibody.SphericalConstraint

Multibody.SphericalSpherical

Multibody.Spring

Multibody.SpringDamperParallel

Multibody.Torque

Multibody.Universal

Multibody.UniversalSpherical

Multibody.World

Multibody.WorldForce

Multibody.WorldTorque

Multibody.absolute_rotation

Multibody.axes_rotations

Multibody.axis_rotation

Multibody.connect_orientation

Multibody.get_frame

Multibody.get_rot

Multibody.get_trans

Multibody.get_w

Multibody.gravity_acceleration

Multibody.ori

Multibody.planar_rotation

Multibody.point_to_point

Multibody.render

Multibody.render!

Multibody.resolve1

Multibody.resolve2

Multibody.traj5

Multibody.urdf2multibody

Multibody.Frame — Function

Frame(; name)

Frame is the fundamental 3D connector in the multibody library. Most components have one or several Frame connectors that can be connected together.

The Frame connector has internal variables for

r_0: The position vector from the world frame to the frame origin, resolved in the world frame

f: The cut force resolved in the connector frame

tau: The cut torque resolved in the connector frame

Depending on usage, also rotation and rotational velocity variables.

Multibody.PlanarMechanics.PartialTwoFrames — Constant

PartialTwoFrames(;name)

Partial model with two frames

Connectors:

- `frame_a` [Frame](@ref) Coordinate system fixed to the component with one cut-force and cut-torque
- `frame_b` [Frame](@ref) Coordinate system fixed to the component with one cut-force and cut-torque

Multibody.PlanarMechanics.ZeroPosition — Constant

ZeroPosition(;name)

Set zero position vector and orientation object of frame_resolve

Connectors:

- `frame_resolve` [FrameResolve](@ref) Coordinate system fixed to the component with one cut-force and cut-torque

Multibody.PlanarMechanics.Frame — Function

Frame(;name)

Coordinate system (2-dim.) fixed to the component with one cut-force and cut-torque

Variables:

x: [m] x position

y: [m] y position

phi: [rad] rotation angle (counterclockwise)

fx: [N] force in the x direction

fy: [N] force in the y direction

tau: [N.m] torque (clockwise)

Parameters:

render: [Bool] Render the joint in animations

length: [m] Length of each axis in animations

radius: [m] Radius of each axis in animations