Sensors

A sensor is an object that translates quantities in the mechanical domain into causal signals which can interact with causal components from ModelingToolkitStandardLibrary.Blocks, such as control systems etc.

Docstrings

Multibody.PlanarMechanics.PartialAbsoluteBaseSensor

Multibody.PlanarMechanics.PartialAbsoluteSensor

Multibody.PlanarMechanics.PartialRelativeBaseSensor

Multibody.PlanarMechanics.PartialRelativeSensor

Multibody.CutForce

Multibody.CutTorque

Multibody.PartialCutForceBaseSensor

Multibody.PlanarMechanics.AbsolutePosition

Multibody.PlanarMechanics.BasicAbsolutePosition

Multibody.PlanarMechanics.BasicRelativePosition

Multibody.PlanarMechanics.RelativePosition

Multibody.Power

Multibody.CutForce — Method

BasicCutForce(; name, resolve_frame)

Basic sensor to measure cut force vector. Contains a connector of type Blocks.RealOutput with name force.

resolve_frame: The frame in which the cut force and cut torque are resolved. Default is :frame_a, options include :frame_a and :world.

Multibody.CutTorque — Method

CutTorque(; name, resolve_frame)

Basic sensor to measure cut torque vector. Contains a connector of type Blocks.RealOutput with name torque.

resolve_frame: The frame in which the cut force and cut torque are resolved. Default is :frame_a, options include :frame_a and :world.

Multibody.PartialCutForceBaseSensor — Method

PartialCutForceBaseSensor(; name, resolve_frame = :frame_a)

resolve_frame: The frame in which the cut force and cut torque are resolved. Default is :frame_a, options include :frame_a and :world.

Multibody.Power — Method

Power(; name)

A sensor measuring mechanical power transmitted from frame_a to frame_b.

Connectors:

power of type RealOutput.

Multibody.PlanarMechanics.PartialAbsoluteBaseSensor — Constant

PartialAbsoluteBaseSensor(;name)

Partial absolute sensor models for sensors defined by equations (frame_resolve must be connected exactly once)

Connectors:

frame_a: 2-dim. Coordinate system from which kinematic quantities are measured

frame_resolve: 2-dim. Coordinate system in which vector is optionally resolved

Multibody.PlanarMechanics.PartialAbsoluteSensor — Constant

PartialAbsoluteSensor(;name)

Partial absolute sensor model for sensors defined by components

Connectors:

`frame: 2-dim. Coordinate system

Multibody.PlanarMechanics.PartialRelativeBaseSensor — Constant

PartialRelativeBaseSensor(;name)

Partial relative sensor models for sensors defined by equations (frame_resolve must be connected exactly once)

Connectors:

frame_a:

frame_b:

frame_resolve:

Multibody.PlanarMechanics.PartialRelativeSensor — Constant

PartialRelativeSensor(;name)

Partial relative sensor model for sensors defined by components

Connectors:

frame_a: Coordinate system a

frame_b: Coordinate system b

Multibody.PlanarMechanics.AbsolutePosition — Method

AbsolutePosition(;name, resolve_in_frame = :frame_a)

Measure absolute position and orientation of the origin of frame connector

Connectors:

x: [m] x-position

y: [m] y-position

phi: [rad] rotation angle (counterclockwise)

Parameters:

resolve_in_frame: Frame in which output x, y, phi is resolved (1: :world, 2: :framea, 3: :frameresolve)

Multibody.PlanarMechanics.BasicAbsolutePosition — Method

BasicAbsolutePosition(;name, resolve_in_frame = :frame_a)

Measure absolute position and orientation (same as Sensors.AbsolutePosition, but frame_resolve is not conditional and must be connected).

Connectors:

x: [m] x-position

y: [m] y-position

phi: [rad] rotation angle (counterclockwise)

frame_a: Coordinate system a

frame_resolve: 2-dim. Coordinate system in which vector is optionally resolved

Parameters:

resolve_in_frame: Frame in which output x, y, phi r is resolved (1: :world, 2: :framea, 3: :frameresolve)

Multibody.PlanarMechanics.BasicRelativePosition — Method

BasicRelativePosition(; name, resolve_in_frame = :frame_a)

Measure relative position and orientation between the origins of two frame connectors

Connectors:

rel_x: [m] Relative x-position

rel_y: [m] Relative y-position

rel_phi: [rad] Relative rotation angle (counterclockwise)

frame_a: Coordinate system a

frame_b: Coordinate system b

frame_resolve:

Parameters:

- `resolve_in_frame`: Frame in which output x, y, phi is resolved (1: :world, 2: :frame_a, 3: frame_b 4: :frame_resolve)

Multibody.PlanarMechanics.RelativePosition — Method

RelativePosition(; name, resolve_in_frame = :frame_a)

Measure relative position and orientation between the origins of two frame connectors

Connectors:

- `rel_x`: [m] Relative x-position
- `re_y`: [m] Relative y-position
- `rel_phi`: [rad] Relative rotation angle (counterclockwise)
- `frame_a`: Coordinate system a
- `frame_b`: Coordinate system b

Parameters:

- `resolve_in_frame`: Frame in which output x, y, phi is resolved (1: :world, 2: :frame_a, 3: frame_b 4: :frame_resolve)