Trajectory_planning

Two methods of planning trajectories are available

point_to_point: Generate a minimum-time point-to-point trajectory with specified start and endpoints, not exceeding specified speed and acceleration limits.

traj5: Generate a 5:th order polynomial trajectory with specified start and end points. Additionally allows specification of start and end values for velocity and acceleration.

Components that make use of these trajectory generators is provided:

KinematicPTP

Kinematic5

These both have output connectors of type RealOutput called q, qd, qdd for positions, velocities and accelerations.

See Industrial robot for an example making use of the point_to_point planner.

Docstrings

Multibody.BodyBox

Multibody.BodyCylinder

Multibody.Planar

Multibody.PlanarMechanics.BodyShape

Multibody.PlanarMechanics.Damper

Multibody.PlanarMechanics.DifferentialGear

Multibody.PlanarMechanics.Fixed

Multibody.PlanarMechanics.FixedTranslation

Multibody.PlanarMechanics.IdealPlanetary

Multibody.PlanarMechanics.PartialAbsoluteBaseSensor

Multibody.PlanarMechanics.PartialAbsoluteSensor

Multibody.PlanarMechanics.PartialRelativeBaseSensor

Multibody.PlanarMechanics.PartialRelativeSensor

Multibody.PlanarMechanics.PartialTwoFrames

Multibody.PlanarMechanics.SimpleWheel

Multibody.PlanarMechanics.Spring

Multibody.PlanarMechanics.SpringDamper

Multibody.PlanarMechanics.ZeroPosition

Multibody.world

Multibody.RotationMatrix

Multibody.AccSensor

Multibody.AxisControlBus

Multibody.AxisType2

Multibody.BasicTorque

Multibody.Body

Multibody.BodyShape

Multibody.CutForce

Multibody.CutTorque

Multibody.Damper

Multibody.Fixed

Multibody.FixedRotation

Multibody.FixedTranslation

Multibody.Force

Multibody.Frame

Multibody.FreeMotion

Multibody.GearConstraint

Multibody.JointRRR

Multibody.JointUSR

Multibody.Kinematic5

Multibody.KinematicPTP

Multibody.NumRotationMatrix

Multibody.PartialCutForceBaseSensor

Multibody.PathPlanning1

Multibody.PathToAxisControlBus

Multibody.PlanarMechanics.AbsolutePosition

Multibody.PlanarMechanics.BasicAbsolutePosition

Multibody.PlanarMechanics.BasicRelativePosition

Multibody.PlanarMechanics.Body

Multibody.PlanarMechanics.Frame

Multibody.PlanarMechanics.Prismatic

Multibody.PlanarMechanics.RelativePosition

Multibody.PlanarMechanics.Revolute

Multibody.PlanarMechanics.SlipBasedWheelJoint

Multibody.PlanarMechanics.limit_S_form

Multibody.PlanarMechanics.limit_S_triple

Multibody.Pose

Multibody.Position

Multibody.Power

Multibody.Prismatic

Multibody.PrismaticConstraint

Multibody.RealPassThrough

Multibody.Revolute

Multibody.RevolutePlanarLoopConstraint

Multibody.RollingConstraintVerticalWheel

Multibody.RollingWheel

Multibody.RollingWheelJoint

Multibody.RollingWheelSet

Multibody.RollingWheelSetJoint

Multibody.Rope

Multibody.SlipWheelJoint

Multibody.SlippingWheel

Multibody.Spherical

Multibody.SphericalConstraint

Multibody.SphericalSpherical

Multibody.Spring

Multibody.SpringDamperParallel

Multibody.Torque

Multibody.Universal

Multibody.UniversalSpherical

Multibody.World

Multibody.WorldForce

Multibody.WorldTorque

Multibody.absolute_rotation

Multibody.axes_rotations

Multibody.axis_rotation

Multibody.connect_orientation

Multibody.get_frame

Multibody.get_rot

Multibody.get_trans

Multibody.get_w

Multibody.gravity_acceleration

Multibody.ori

Multibody.planar_rotation

Multibody.point_to_point

Multibody.render

Multibody.render!

Multibody.resolve1

Multibody.resolve2

Multibody.traj5

Multibody.urdf2multibody

Multibody.Kinematic5 — Method

Kinematic5(; time, name, q0 = 0, q1 = 1, qd0 = 0, qd1 = 0, qdd0 = 0, qdd1 = 0)

A component emitting a 5:th order polynomial trajectory created using traj5. traj5 is a simple trajectory planner that plans a 5:th order polynomial trajectory between two points, subject to specified boundary conditions on the position, velocity and acceleration.

Arguments

time: Time vector, e.g., 0:0.01:10

name: Name of the component

Outputs

q: Position

qd: Velocity

qdd: Acceleration

Multibody.KinematicPTP — Method

KinematicPTP(; time, name, q0 = 0, q1 = 1, qd_max=1, qdd_max=1)

A component emitting a trajectory created by the point_to_point trajectory generator.

Arguments

time: Time vector, e.g., 0:0.01:10

name: Name of the component

q0: Initial position

q1: Final position

qd_max: Maximum velocity

qdd_max: Maximum acceleration

Outputs

q: Position

qd: Velocity

qdd: Acceleration

See also Kinematic5.

Multibody.PathPlanning1 — Method

Generate reference angles for specified kinematic movement

Multibody.PathToAxisControlBus — Method

Map path planning to one axis control bus

Multibody.RealPassThrough — Method

RealPassThrough(; name)

Pass a Real signal through without modification

Connectors

input

output

Multibody.traj5 — Method

q, qd, qdd = traj5(t; q0, q1, q̇0 = zero(q0), q̇1 = zero(q0), q̈0 = zero(q0), q̈1 = zero(q0))

Generate a 5:th order polynomial trajectory with specified end points, vels and accs.

See also point_to_point and Kinematic5.

Multibody.point_to_point — Method

q,qd,qdd,t_end = point_to_point(time; q0 = 0.0, q1 = 1.0, t0 = 0, qd_max = 1, qdd_max = 1)

Generate a minimum-time point-to-point trajectory with specified start and endpoints, not exceeding specified speed and acceleration limits.

The trajectory produced by this function will typically exhibit piecewise constant accleration, piecewise linear velocity and piecewise quadratic position curves.

If a vector of time points is provided, the function returns matrices q,qd,qdd of size (length(time), n_dims). If a scalar time point is provided, the function returns q,qd,qdd as vectors with the specified dimension (same dimension as q0). t_end is the time at which the trajectory will reach the specified end position.

Arguments:

time: A scalar or a vector of time points.

q0: Initial coordinate, may be a scalar or a vector.

q1: End coordinate

t0: Tiem at which the motion starts. If time contains time points before t0, the trajectory will stand still at q0 until time reaches t0.

qd_max: Maximum allowed speed.

qdd_max: Maximum allowed acceleration.

See also KinematicPTP and traj5.