Forces

Multibody.BasicTorque

Multibody.Damper

Multibody.Force

Multibody.Spring

Multibody.SpringDamperParallel

Multibody.Torque

Multibody.WorldForce

Multibody.WorldTorque

Multibody.BasicTorque — Method

BasicTorque(; name, resolve_frame = :world)

Low-level torque component used to build Torque

Multibody.Damper — Method

Damper(; d, name, kwargs)

Linear damper acting as line force between frame_a and frame_b. A force f is exerted on the origin of frame_b and with opposite sign on the origin of frame_a along the line from the origin of frame_a to the origin of frame_b according to the equation:

\[f = d D(s)\]

where d is the (viscous) damping parameter, s is the distance between the origin of frame_a and the origin of frame_b and D(s) is the time derivative of s.

Arguments:

d: Damping coefficient

Rendering

radius = 0.1: Radius of damper when rendered

length_fraction = 0.2: Fraction of the length of the damper that is rendered

color = [0.5, 0.5, 0.5, 1]: Color of the damper when rendered

See also SpringDamperParallel

Multibody.Force — Method

Force(; name, resolve_frame = :frame_b)

Force acting between two frames, defined by 3 input signals and resolved in frame world, frame_a, frame_b (default)

Connectors:

frame_a

frame_b

force: Of type Blocks.RealInput(3). x-, y-, z-coordinates of force resolved in frame defined by resolve_frame.

Keyword arguments:

resolve_frame: The frame in which the cut force and cut torque are resolved. Default is :frame_b, options include :frame_a and :world.

Multibody.Spring — Method

Spring(; c, name, m = 0, lengthfraction = 0.5, s_unstretched = 0, kwargs)

Linear spring acting as line force between frame_a and frame_b. A force f is exerted on the origin of frame_b and with opposite sign on the origin of frame_a along the line from the origin of frame_a to the origin of frame_b according to the equation:

\[f = c s\]

where c is the spring stiffness parameter, s is the distance between the origin of frame_a and the origin of frame_b.

Optionally, the mass of the spring is taken into account by a point mass located on the line between frame_a and frame_b (default: middle of the line). If the spring mass is zero, the additional equations to handle the mass are removed.

Arguments:

c: Spring stiffness

m: Mass of the spring (can be zero)

lengthfraction: Location of spring mass with respect to frame_a as a fraction of the distance from frame_a to frame_b (=0: at frame_a; =1: at frame_b)

s_unstretched: Length of the spring when it is unstretched

kwargs: are passed to LineForceWithMass

Rendering

num_windings = 6: Number of windings of the coil when rendered

color = [0,0,1,1]: Color of the spring when rendered

radius = 0.1: Radius of spring when rendered

N = 200: Number of points in mesh when rendered. Rendering time can be reduced somewhat by reducing this number.

See also SpringDamperParallel

Multibody.SpringDamperParallel — Method

SpringDamperParallel(; name, c, d, s_unstretched)

Linear spring and linear damper in parallel acting as line force between frame_a and frame_b. A force f is exerted on the origin of frame_b and with opposite sign on the origin of frame_a along the line from the origin of frame_a to the origin of frame_b according to the equation:

\[f = c (s - s_{unstretched}) + d \cdot D(s)\]

where c, s_unstretched and d are parameters, s is the distance between the origin of frame_a and the origin of frame_b and D(s) is the time derivative of s.

Multibody.Torque — Method

Torque(; name, resolve_frame = :frame_b)

Torque acting between two frames, defined by 3 input signals and resolved in frame world, frame_a, frame_b (default)

Connectors:

frame_a

frame_b

torque: Of type Blocks.RealInput(3). x-, y-, z-coordinates of torque resolved in frame defined by resolve_frame.

Keyword arguments:

resolve_frame: The frame in which the cut force and cut torque are resolved. Default is :frame_b, options include :frame_a and :world.

Multibody.WorldForce — Method

WorldForce(; name, resolve_frame = :world)

External force acting at frame_b, defined by 3 input signals and resolved in frame :world or :frame_b.

Connectors:

frame_b: Frame at which the force is acting

force: Of type Blocks.RealInput(3). x-, y-, z-coordinates of force resolved in frame defined by resolve_frame.

Rendering options

scale = 0.1: scaling factor for the force [m/N]

color = [0,1,0,0.5]: color of the force arrow in rendering

radius = 0.05: radius of the force arrow in rendering

Multibody.WorldTorque — Method

WorldTorque(; name, resolve_frame = :world)

External torque acting at frame_b, defined by 3 input signals and resolved in frame :world or :frame_b.

Connectors:

frame_b: Frame at which the torque is acting

torque: Of type Blocks.RealInput(3). x-, y-, z-coordinates of torque resolved in frame defined by resolve_frame.

Rendering options

scale = 0.1: scaling factor for the force [m/N]

color = [0,1,0,0.5]: color of the force arrow in rendering

radius = 0.05: radius of the force arrow in rendering